Update TenancyPerSchema.cs
Oskar.   If you guys decide you would like to move forward with introducing a full DocumentStoreFactory would you please keep in mind that my scenario doesn't directly match your example.  I'm not really looking to change the schema per tenant, but rather ANY number of configuration values per store creation.   The only change to your code was to allow the consumer/dev to decide on what the key field is on the document store and to allow a complex object to be passed to the configure method.   Again, I appreciate the assist with this.