Added example implementation of Passage Of Time and To-Do List patterns
Added example implementation of Passage Of Time and To-Do List patterns
Read more in:
- Pasage Of Time by @mathiasverraes  - https://verraes.net/2019/05/patterns-for-decoupling-distsys-passage-of-time-event/
- The To-Do List Pattern by @yreynhout - https://blog.bittacklr.be/the-to-do-list-pattern.html

I'm using Quartz.NET to trigger passage of time events.

Note: the current implementation is a bit "old school" and needs to be refactored; that's why I'm orchestrating the passage of time in the application layer instead of pushing events to the domain (as in the original article).